Extract OpenAI image provider adapter
